Update to 1.3 (Completed) - Added Output Text Area  - Provides instructions and usage info  - Provides error output, commands used etc.  - Unlimited capacity  - Exportable to log files - Added cookie support  - Added enable cookies checkbox  - Added regenerate button  - Scans downloads folder for exported cookie files  - Creates single cookies file  - Allows for authentication for member videos etc. - Improved download path determination with three different methods - Command objects now created via arg arraylist, not array - Added FileIO class for I/O helper methods - Added intermittent dot printing during long operations (liveness) - Added version check for youtube-dl.exe - Added URL refining - extracts ID - Improved moveToDownloads  - Fixed issue where it broke when downloads path had space(s)  - Added handling for when diff between dir before/after is not 1 - Downloaded files have their path and readable size printed - Removed invisible extra menu JComponents (used Navigator feature) - Add tutorial link to menu - Added quoting to URLs to improve stability - Improved default format name - Updated preview screenshot, Preserved previous. - Added screenshots of member video process - Updated copy script to account for new GDrive folder 'MyOwn' - Compiled on JDK 8 to make running easier, only JRE needed not JDK - Increase window size - Improved colour scheme

+++ Project/src/main/FileIO.java
@@ -2,12 +2,13 @@ package main;
 
 import java.io.File;
 import java.io.IOException;
-import java.net.CookieStore;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
+import java.nio.file.LinkOption;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardOpenOption;
+import java.text.DecimalFormat;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -25,10 +26,22 @@ public class FileIO {
     // Cookies file path
     public static final String cookieFP = "cookies.txt";
 
+    // Substring used to determine whether a text file is an exported cookie file
+    private final String validCookieSS = "# Netscape HTTP Cookie File";
+
     /**
-     * Generate cookie file
+     * Generate cookie file if cookies are enabled
+     *
+     * @param regen Delete and regenerate option
      */
-    public void genCookieFile() {
+    public void genCookieFile(boolean regen) {
+
+        // If cookies are not enabled
+        if (!GUI.gui.isCookieFileEnabled()) {
+
+            // Do not process further
+            return;
+        }
 
         // Space
         outputln("");
@@ -37,14 +50,27 @@ public class FileIO {
         if (isValidPath(cookieFP)) {
 
             // Notify
-            outputln("Previous '" + cookieFP
-                    + "' file was found next to exe! "
-                    + "This file will be used.");
-            outputln("If you wish to regenerate the cookies file, "
-                    + "delete this file and try enabling cookies again.");
+            outputln("Previous '" + cookieFP + "' file was found.");
 
-            // Do not process further
-            return;
+            // If regeneration is wanted
+            if (regen) {
+
+                // Delete cookie file
+                deleteFile(cookieFP);
+
+                // Notify
+                outputln("Deleted previous cookies file. Regenerating...");
+
+            } else {
+
+                // Else if regeneration is not wanted
+                outputln("This cookies file will be used. To make "
+                        + "a new cookies file, use the regenerate button.");
+
+                // Do not process further
+                return;
+            }
+
         } else {
 
             // Else if no cookies file exists,
@@ -57,10 +83,10 @@ public class FileIO {
 
         // Notify about number of downloads files
         int numF = paths.length;
-        outputln("Detected " + numF + " files in downloads folder.");
+        outputln("Detected " + numF + " files in Downloads folder.");
         int fLimit = 30;
         if (numF > fLimit) {
-            outputln("Since this program scans all downloads folder files, "
+            outputln("Since this program scans all Downloads folder files, "
                     + "it would be a good idea to "
                     + "reduce the number of files present "
                     + "(to below " + fLimit + ")");
@@ -68,9 +94,6 @@ public class FileIO {
 
         // Cookies file string
         String cookiesFS = "";
-
-        // Valid cookie file substring
-        String cookieSubstr = "# Netscape HTTP Cookie File";
 
         // For each filepath in paths
         for (String fp : paths) {
@@ -87,26 +110,14 @@ public class FileIO {
                 // If file is definitely a text file
                 if (file.endsWith(".txt")) {
 
-                    // Read in contents
-                    String curContents = getFileAsString(fp);
+                    // Process text file
+                    String contents = processTxtFileinDwl(fp, file);
 
-                    // If seems like cookie file
-                    if (curContents.contains(cookieSubstr)) {
-
-                        // Notify
-                        outputln("Retrieved data from "
-                                + "valid exported cookie file: " + file);
+                    // If contents is not null
+                    if (contents != null) {
 
                         // Add to cookie file string
-                        cookiesFS += curContents;
-                    } else {
-
-                        // If non-valid cookie text file was found, notify
-                        outputln("Found text file (" + file
-                                + "), but was not "
-                                + "valid exported cookie file.");
-                        outputln("(i.e. didn't contain: '"
-                                + cookieSubstr + "')");
+                        cookiesFS += contents;
                     }
                 }
             }
@@ -115,7 +126,7 @@ public class FileIO {
         // If cookie file string is empty
         if (cookiesFS.isEmpty()) {
             outputln("No exported cookie text files were "
-                    + "found in the downloads folder.");
+                    + "found in the Downloads folder.");
             outputln("Cookies file could not be generated.");
             return;
         }
@@ -128,7 +139,7 @@ public class FileIO {
     }
 
     /**
-     * Move the downloaded video to the downloads folder
+     * Move the downloaded video to the Downloads folder
      *
      * @param dirBefore
      * @return
@@ -147,29 +158,40 @@ public class FileIO {
             dirAL.remove(curFile);
         }
 
-        // Arguments
-        String[] args = new String[2];
+        // Move command arguments
+        ArrayList<String> movArgs = new ArrayList<>();
 
         // If one file remains
         if (dirAL.size() == 1) {
 
             // Get the string remaining - the new file
+            // This must be the downloaded media file
             String newFile = dirAL.get(0);
             newFile = newFile.replace(".\\", "");
-            newFile = "\"" + newFile + "\"";
 
             // Move file to downloads
             String prog = "move";
-            args[0] = newFile;
-            args[1] = downloadsPath;
-            Command comm = new Command(prog, args);
+            movArgs.add(Code.quoteS(newFile));
+            movArgs.add(Code.quoteS(downloadsPath));
+            Command comm = new Command(prog, movArgs);
             comm.run();
-        }
 
-        // Deduce final path and return
-        args[0] = args[0].replace("\"", "");
-        String finalPath = args[1] + "\\" + args[0];
-        return finalPath;
+            // Deduce final path and return
+            String finalPath = movArgs.get(1) + "\\";
+            finalPath += movArgs.get(0);
+            finalPath = finalPath.replace("\"", "");
+            return finalPath;
+        } else {
+
+            // Else, there is no new file detected
+            outputln("Error: Could not move file to downloads");
+            outputln("Debug Info:");
+            outputln("DirBefore: " + Arrays.toString(dirBefore));
+            outputln("DirAfter: " + Arrays.toString(dirAfter));
+
+            // Return where file is
+            return "next to exe";
+        }
     }
 
     /**
@@ -219,6 +241,76 @@ public class FileIO {
     }
 
     /**
+     * Process a text file in the Downloads folder
+     *
+     * @param fp The file path
+     * @param file The file's name and extension
+     * @return Contents if valid cookie file, otherwise null
+     */
+    private String processTxtFileinDwl(String fp, String file) {
+
+        // Read in contents
+        String curContents = getFileAsString(fp);
+
+        // If seems like cookie file
+        if (curContents.contains(validCookieSS)) {
+
+            // Notify
+            outputln("Retrieved data from valid exported cookie file: "
+                    + file);
+
+            // Return contents for adding to cookies file
+            return curContents;
+        } else {
+
+            // If non-valid cookie text file was found, notify
+            outputln("Found text file (" + file + "), but was not valid "
+                    + "exported cookie file.");
+            outputln("(i.e. didn't contain: '"
+                    + validCookieSS + "')");
+
+            // Return null
+            return null;
+        }
+    }
+
+    /**
+     * Get readable size of a given file
+     *
+     * @param fp File path
+     * @return
+     */
+    public static String getReadableFileSize(String fp) {
+
+        // Get raw file size
+        long size = 0;
+        try {
+            size = Files.size(Paths.get(fp));
+        } catch (IOException ex) {
+            outputerr(ex);
+        }
+
+        // If size is 0
+        if (size <= 0) {
+
+            // Return 0 bytes
+            return "0 B";
+        }
+
+        // Add size
+        int dg = (int) (Math.log10(size) / Math.log10(1024));
+        String rs;
+        rs = new DecimalFormat("#,##0.#").format(size / Math.pow(1024, dg));
+
+        // Add unit
+        String[] units = {"B", "kB", "MB", "GB", "TB"};
+        rs += " " + units[dg];
+
+        // Return readable size
+        return rs;
+    }
+
+    /**
      * Get contents of a file as a string
      *
      * @param fpS File path string
@@ -265,12 +357,40 @@ public class FileIO {
     }
 
     /**
+     * Hide a given file
+     *
+     * @param fp File path
+     */
+    public static void hideFile(String fp) {
+        try {
+            
+            // Get path
+            Path path = Paths.get(fp);
+            
+            // Set hidden attribute to true
+            LinkOption lo = LinkOption.NOFOLLOW_LINKS;
+            Files.setAttribute(path, "dos:hidden", true, lo);
+        } catch (IOException e) {
+            Code.outputerr(e);
+        }
+    }
+
+    /**
      * Delete file at given file path
      *
      * @param fp
      * @return True if successful
      */
     public boolean deleteFile(String fp) {
+
+        // If file does not exist
+        if (!isValidPath(fp)) {
+
+            // Return false as cannot delete
+            return false;
+        }
+
+        // Return true if file was deleted
         return new File(fp).delete();
     }
 
@@ -278,10 +398,9 @@ public class FileIO {
      * Return true if path is valid file or folder
      *
      * @param path
-     * @return
+     * @return True if valid
      */
     public boolean isValidPath(String path) {
         return (new File(path)).exists();
     }
-
 }
